Locations now uses doubles (double x, double y)

+++ src/Locations.java
@@ -1,22 +1,22 @@
 public class Locations {
-  private int x;
-  private int y;
-  public Locations(int x, int y){
+  private double x;
+  private double y;
+  public Locations(double x, double y){
     this.x = x;
     this.y = y;
   }
 
-  public int getX(){
+  public double getX(){
     return this.x;
   }
-  public int getY(){
+  public double getY(){
     return this.y;
   }
 
-  public void setX(int x){
+  public void setX(double x){
     this.x = x;
   }
-  public void setY(int y){
+  public void setY(double y){
     this.y = y;
   }
